refactor : NumberGenerator public 클래스를 Balls 의 static 클래스로 변경
Balls 에서만 사용하기 때문에 Balls 의 inner class 로 사용
